// src/lib/gen/carbon/v1/common_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file carbon/v1/common.proto (package carbon.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file carbon/v1/common.proto.
 */
export const file_carbon_v1_common: GenFile = /*@__PURE__*/
  fileDesc("ChZjYXJib24vdjEvY29tbW9uLnByb3RvEgljYXJib24udjEiYAoIQ2l0YXRpb24SCgoCaWQYASABKAkSDQoFdGl0bGUYAiABKAkSDwoHYXV0aG9ycxgDIAMoCRIMCgR5ZWFyGAQgASgFEg0KBXZlbnVlGAUgASgJEgsKA3VybBgGIAEoCSJeCgtNZXRyaWNSYW5nZRILCgNsb3cYASABKAESDAoEaGlnaBgCIAEoARIMCgR1bml0GAMgASgJEhIKCnllYXJfYmFzaXMYBCABKAUSEgoKc291cmNlX3JlZhgFIAEoCWIGcHJvdG8z");

/**
 * @generated from message carbon.v1.Citation
 */
export type Citation = Message<"carbon.v1.Citation"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * @generated from field: repeated string authors = 3;
   */
  authors: string[];

  /**
   * @generated from field: int32 year = 4;
   */
  year: number;

  /**
   * @generated from field: string venue = 5;
   */
  venue: string;

  /**
   * @generated from field: string url = 6;
   */
  url: string;
};

/**
 * Describes the message carbon.v1.Citation.
 * Use `create(CitationSchema)` to create a new message.
 */
export const CitationSchema: GenMessage<Citation> = /*@__PURE__*/
  messageDesc(file_carbon_v1_common, 0);

/**
 * @generated from message carbon.v1.MetricRange
 */
export type MetricRange = Message<"carbon.v1.MetricRange"> & {
  /**
   * @generated from field: double low = 1;
   */
  low: number;

  /**
   * @generated from field: double high = 2;
   */
  high: number;

  /**
   * validated against UNIT_ALLOWLIST in seed loader
   *
   * @generated from field: string unit = 3;
   */
  unit: string;

  /**
   * @generated from field: int32 year_basis = 4;
   */
  yearBasis: number;

  /**
   * -> Citation.id, required
   *
   * @generated from field: string source_ref = 5;
   */
  sourceRef: string;
};

/**
 * Describes the message carbon.v1.MetricRange.
 * Use `create(MetricRangeSchema)` to create a new message.
 */
export const MetricRangeSchema: GenMessage<MetricRange> = /*@__PURE__*/
  messageDesc(file_carbon_v1_common, 1);